+++ index.html
@@ -16,8 +16,9 @@
     <main id="main-container">
       <div id="ledger-view" class="view active-view">
         <div id="input-form-container"></div>
+        <div id="total-amount-container"></div>
 
-        <div class="total-amount">
+        <!-- <div class="total-amount">
           <div class="total-amount-text" id="total-amount-text">전체 내역</div>
           <div class="total-amount-value" id="total-amount-value">
             <div
@@ -35,7 +36,7 @@
               지출
             </div>
           </div>
-        </div>
+        </div> -->
 
         <section class="entry-list" id="entry-list"></section>
       </div>
